// frontend/generated/models/SlotDto.ts
export interface SlotDto {
    /**
     * 
     * @type {number}
     * @memberof SlotDto
     */
    slotId?: number;
    /**
     * 
     * @type {number}
     * @memberof SlotDto
     */
    position?: number;
    /**
     * 
     * @type {string}
     * @memberof SlotDto
     */
    state?: SlotDtoStateEnum;
    /**
     * 
     * @type {boolean}
     * @memberof SlotDto
     */
    isVisible?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof SlotDto
     */
    openedAt?: Date;
    /**
     * 
     * @type {ToyDto}
     * @memberof SlotDto
     */
    toy?: ToyDto;
    /**
     * 
     * @type {number}
     * @memberof SlotDto
     */
    setId?: number;
    /**
     * 
     * @type {Date}
     * @memberof SlotDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof SlotDto
     */
    updatedAt?: Date;
    /**
     * 
     * @type {VideoDto}
     * @memberof SlotDto
     */
    video?: VideoDto;
}


/**
 * @export
 */
export const SlotDtoStateEnum = {
    Opened: 'OPENED',
    Available: 'AVAILABLE',
    Reserved: 'RESERVED'
} as const;
export type SlotDtoStateEnum = typeof SlotDtoStateEnum[keyof typeof SlotDtoStateEnum];